pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.5",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "rich>=14.1.0",
]
